+++ orbitSimulator/orbitSimulator/position.h
@@ -29,38 +29,38 @@ public:
    
 	// constructors
 	Position() : x(0.0), y(0.0) {}
-	Position(double x, double y);
+	Position(float x, float y);
 	Position(const Position & pt) : x(pt.x), y(pt.y) {}
 	Position& operator = (const Position& pt);
 
 	// getters
-	double getMetersX()       const { return x;                    }
-	double getMetersY()       const { return y;                    }
-	double getPixelsX()       const { return x / metersFromPixels; }
-	double getPixelsY()       const { return y / metersFromPixels; }
+	float getMetersX()       const { return x;                    }
+	float getMetersY()       const { return y;                    }
+	float getPixelsX()       const { return x / metersFromPixels; }
+	float getPixelsY()       const { return y / metersFromPixels; }
 
 	// setters
-	void setMeters(double xMeters, double yMeters) {x = xMeters; y = yMeters; }
-	void setMetersX(double xMeters)       { x = xMeters;           }
-	void setMetersY(double yMeters)       { y = yMeters;           }
-	void setPixelsX(double xPixels)       { x = xPixels * metersFromPixels;          }
-	void setPixelsY(double yPixels)       { y = yPixels * metersFromPixels;          }
-	void addMetersX(double dxMeters)      { setMetersX(getMetersX() + dxMeters);     }
-	void addMetersY(double dyMeters)      { setMetersY(getMetersY() + dyMeters);     }
-	void addPixelsX(double dxPixels)      { setPixelsX(getPixelsX() + dxPixels);     }
-	void addPixelsY(double dyPixels)      { setPixelsY(getPixelsY() + dyPixels);     }
+	void setMeters(float xMeters, float yMeters) {x = xMeters; y = yMeters; }
+	void setMetersX(float xMeters)       { x = xMeters;           }
+	void setMetersY(float yMeters)       { y = yMeters;           }
+	void setPixelsX(float xPixels)       { x = xPixels * metersFromPixels;          }
+	void setPixelsY(float yPixels)       { y = yPixels * metersFromPixels;          }
+	void addMetersX(float dxMeters)      { setMetersX(getMetersX() + dxMeters);     }
+	void addMetersY(float dyMeters)      { setMetersY(getMetersY() + dyMeters);     }
+	void addPixelsX(float dxPixels)      { setPixelsX(getPixelsX() + dxPixels);     }
+	void addPixelsY(float dyPixels)      { setPixelsY(getPixelsY() + dyPixels);     }
 
 	// deal with the ratio of meters to pixels
-	void setZoom(double metersFromPixels)
+	void setZoom(float metersFromPixels)
 	{
 	   this->metersFromPixels = metersFromPixels;
 	}
-	double convertToMeters(double pixels) { return pixels * metersFromPixels; }
-	double getZoom() const { return metersFromPixels; }
+	float convertToMeters(float pixels) { return pixels * metersFromPixels; }
+	float getZoom() const { return metersFromPixels; }
 
 private:
-	double x;                 // horizontal position
-	double y;                 // vertical position
+	float x;                 // horizontal position
+	float y;                 // vertical position
 	static double metersFromPixels;
 };
 
@@ -68,7 +68,7 @@ private:
  * COMPUTE DISTANCE
  * Find the distance between two positions
  *********************************************/
-inline double computeDistance(const Position& pos1, const Position& pos2)
+inline float computeDistance(const Position& pos1, const Position& pos2)
 {
    return sqrt((pos1.getMetersX() - pos2.getMetersX()) * (pos1.getMetersX() - pos2.getMetersX()) +
 			   (pos1.getMetersY() - pos2.getMetersY()) * (pos1.getMetersY() - pos2.getMetersY()));
@@ -85,8 +85,8 @@ std::istream & operator >> (std::istream & in,        Position& pt);
  *********************************************/
 struct PT
 {
-	double x;
-	double y;
+	float x;
+	float y;
 };
 
 #endif /* position_h */
